improve: sort out options
